# Conflicts:
#	jenetics.ext/src/main/java/io/jenetics/ext/engine/AdaptiveEngine.java
#	jenetics.ext/src/test/java/io/jenetics/ext/engine/AdaptiveEngineTest.java
#	jenetics/src/main/java/io/jenetics/Phenotype.java
#	jenetics/src/main/java/io/jenetics/engine/Engine.java
#	jenetics/src/main/java/io/jenetics/engine/EvolutionStart.java
#	jenetics/src/main/java/io/jenetics/engine/EvolutionStreamable.java

+++ jenetics/src/test/java/io/jenetics/ProbabilitySelectorTest.java
@@ -74,7 +74,7 @@ public class ProbabilitySelectorTest {
 
 		final double[] reverted = ProbabilitySelector.sortAndRevert(values);
 		for (int i = 0; i < values.length; ++i) {
-			Assert.assertEquals(reverted[i], (double)(values.length - i - 1));
+			Assert.assertEquals(reverted[i], (values.length - i - 1));
 		}
 	}
